--- game-romb-server/src/users/dto/create-user.dto.ts
@@ -1,12 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class CreateUserDto {
-
-    @IsNotEmpty()
-    @IsString()
-    readonly nickname: string;
-
-    @IsNotEmpty()
-    @IsString()
-    readonly password: string;
-}
